+++ R/grade.r
@@ -1,9 +1,81 @@
-#' A grading function.
+#' This function finds unreadable files.
+#'
+#'  A function that finds student submissions with poorly encoded characters
+#' @param submission_dir where the assignments are located
+#' @keywords calcGrades findBadEncodingFiles
+#' @export
+#' @examples
+#' \donttest{
+#' # change paths to *your* paths
+#' submissions <- "extdata/assignment1_submissions/"
+#' findBadEncodingFiles(submissions) # perhaps ask these students to resubmit
+#' }
+findBadEncodingFiles <- function(submission_dir){
+  scripts_to_grade <- list.files(path = submission_dir, 
+                                 recursive = T, 
+                                 pattern = "\\.r$", 
+                                 ignore.case = T)  
+  atLeastOneBadFile <- FALSE
+  for(script in scripts_to_grade) {
+    globalPath <- paste(submission_dir, script, sep = "")   
+    lines <- readLines(globalPath, warn = F)
+    badLines <- lines[!validUTF8(readLines(globalPath, warn = F))]
+    
+    if(length(badLines) > 0) {
+      atLeastOneBadFile <- TRUE
+      cat("======================================================================\n")
+      cat("In ", script, "\n")
+      for(elem in badLines) cat(elem, "\n")
+    }
+  }
+  if(atLeastOneBadFile) 
+    cat("======================================================================\n")
+}
+
+
+#' This function finds files with global file paths.
+#'
+#'  A function that finds student submissions that refer to machine-specific file paths
+#' @param submission_dir where the assignments are located
+#' @keywords calcGrades findGlobalPaths
+#' @export
+#' @examples
+#' \donttest{
+#' # change paths to *your* paths
+#' submissions <- "extdata/assignment1_submissions/"
+#' findGlobalPaths(submissions) # perhaps ask these students to resubmit
+#' }
+findGlobalPaths <- function(submission_dir) {
+  scripts_to_grade <- list.files(path = submission_dir, 
+                                 recursive = T, 
+                                 pattern = "\\.r$", 
+                                 ignore.case = T)
+  
+  atLeastOneBadFile <- FALSE
+  for(script in scripts_to_grade) {
+    globalPath <- paste(submission_dir, script, sep = "")
+    lines <- readLines(globalPath, warn = F)  
+    badLines <- suppressWarnings(lines[grep("^[^#].+[\\\\/]+", lines)])
+    
+    if(length(badLines) > 0) {
+      atLeastOneBadFile <- TRUE
+      cat("======================================================================\n")
+      cat("In ", script, "\n")
+      for(elem in badLines) cat(elem, "\n")
+    }
+  }
+  if(atLeastOneBadFile) 
+    cat("======================================================================\n")
+}
+
+
+#' The grading function.
 #'
 #' This function grades a bunch of R script assignments 
 #' @param submission_dir where the assignments are located
 #' @param your_test_file the path to your testthat test file (e.g. grade_hw1.R)
 #' @param suppress_warnings warning handlers prevent code from being run after they catch something. Suppress this behavior by setting this argument to TRUE.
+#' @param verbose set to true if you want to print the name of the file as it's being ran
 #' @keywords calcGrades
 #' @export
 #' @examples
@@ -13,7 +85,7 @@
 #' my_test_file <- system.file("extdata", "grade_hw1.R", package = "gradeR")
 #' results <- calcGrades(submissions, my_test_file)
 #' }
-calcGrades <- function(submission_dir, your_test_file, suppress_warnings = TRUE){
+calcGrades <- function(submission_dir, your_test_file, suppress_warnings = TRUE, verbose = FALSE){
 
   if(missing(submission_dir) | missing(your_test_file)) 
     stop("the first two arguments are required")
@@ -38,6 +110,8 @@ calcGrades <- function(submission_dir, your_test_file, suppress_warnings = TRUE)
     testEnv <- new.env()
     if(suppress_warnings){
       tryCatch({
+        if(verbose)
+          cat("grading: ", path, "\n")
         suppressWarnings(source(tmp_full_path, testEnv))
       }, error = function(e) {
         
@@ -48,6 +122,8 @@ calcGrades <- function(submission_dir, your_test_file, suppress_warnings = TRUE)
       })
     } else { #not suppressing warnings
       tryCatch({
+        if(verbose)
+          cat("grading: ", path, "\n")
         source(tmp_full_path, testEnv)
       }, error = function(e) {