+++ NEWS.md
@@ -1,3 +1,11 @@
+# gradeR 1.0.5
+
+## Extra things
+
+- added `findGlobalPaths`, which searchs student submissions for the very bad global/machine-specific file paths
+- added `findBadEncodingFiles`, which searchs student submissions for the very bad non-UTF-8 characters
+- added a `verbose` argument to `calcGrades`, so you can see which file is being ran in real time
+
 # gradeR 1.0.4
 
 ## Extra things